v1.4 ys 3차 완료 및 로직 개선  - 리스트 구현 중
