+++ src/main/java/com/example/psychologicaltest/controller/Controller.java
@@ -1,6 +1,8 @@
 package com.example.psychologicaltest.controller;
 
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 
 @org.springframework.stereotype.Controller
 public class Controller {
@@ -12,8 +14,10 @@ public class Controller {
     }
 
     @GetMapping("/ys/ysfinish")
-    public String intoResult() {
+    public String intoResult(@RequestParam int data, Model model) {
 
+        System.out.println(data);
+        model.addAttribute("data", data);
         return "ys/ysfinish";
     }
 }
