+++ src/main/resources/templates/ys/ystest.html
@@ -68,14 +68,15 @@
         <input type="radio" name="mgender" value="male">남자
         <input type="radio" name="mgender" value="female">여자<br/><br/>
 
-        <!--<button id="btn-save" class="btn btn-primary" onclick="js:insertResult()">결과보기</button>-->
         <button type="button" class="btn btn-primary" onclick="js:insertMemberInfo()">결과보기</button>
 
         </form></div>
     </section>
 
+
     <script type="text/javascript" th:src="@{/js/data_ys.js}"></script>
-    <script type="text/javascript" th:src="@{/js/start_ys.js}"></script>
+    <script type="text/javascript" th:src="@{/js/ysTestLogic.js}"></script>
+    <script type="text/javascript" th:src="@{/js/ysResultView.js}"></script>
 
 </div>
 </body>
